requirements.txt
Flask
gunicorn
psycopg2-binary
python-dotenv
requests
beautifulsoup4
tqdm
google-generativeai